Move test function back, fixup metrics and tests
